requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4==4.12.3
python-dotenv
requests
python-telegram-bot==21.3
selenium
